Add additional class to closebrowser rescue
Previously, closebrowser would throw a Selenium::WebDriver::Error::WebDriverError when closebrowser was called on either a non-existent browser or a system that could not bring up a browser.  Unfortunately, this would not throw an error in the at_exit hook in RSpec and cause the whole test to pass!  This fix adds the error and hopefully now, tests will no longer lie when they fail miserably.